feat: get wallet balance

+++ sql/migrations/000002_create_wallets_table.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS wallets(
    wallet_id UUID NOT NULL DEFAULT gen_random_uuid(),
    user_id UUID NOT NULL REFERENCES users(user_id),
-   balance NUMERIC(12, 2) DEFAULT 0,
+   balance BIGINT DEFAULT 0,
    created_at timestamp DEFAULT now()
 );
